#f3175eb clean failed wget
When wget fails (because requested file is not available), it even tough
creates the ouput file specified with -O switch. This makes the further
fallback URL not trying to download or link the expected file because of
the already existing target.
Then on wget failer, the target file must be deleted.