#566131a refactoring XARCH / CROSS_ARCH features
XARCH and CROSS_ARCH may be used indifferently and in most case
overriding any of SYSNAME, HWNAME or ARCH should provide the same
results.

Overall strategy changed to consider in most case ARCH to be the target
architecture. Then:
 - HOST_SYSNAME, HOST_HWNAME and HOST_ARCH are introduced to identify
   architecture of the host running abs.
 - by default, SYSNAME, HWNAME and ARCH are respectively equals to
   HOST_SYSNAME, HOST_HWNAME, HOST_ARCH
 - overriding any of this default obviously trigger some kind of cross
   compilation and involves the selection of an alternate compiler (if
   not explicitely specified) and the selection of the dependencies
   matching the overridden target.
 - when some dependencies for the build itself, using the host
   architecture (and obviouly not include in the dist archive) are
   needed, the HOST_USELIB variable shall be defined to list such
   dependencies.

In some cases some effort have been made to maintain backword
compatibility and CROSS_ARCH & XARCH variable may be used as before this
commit. But some doubts remains and the efforts were limited. In case of
regression, try the following:
   - reconsider USELIB definitions (and all *_USELIB definitions)
     matching the rules given above: USELIBi/NA_USELIB for target deps,
     HOST_USELIB for "local" host deps.
   - override HWNAME to just select an alternate CPU architecture instead of
     defining CROSS_ARCH